+++ CHANGELOG.md
@@ -4,6 +4,8 @@ _Note: Numbers like (\#123) point to closed Pull Requests on the fractal-web rep
 
 - Add support for duplicating workflowtasks (\#1077).
 - Add `fractalJobId` for `/api/admin/v2/accounting/slurm` request (\#1081).
+- Add new users-info admin page (\#1079).
+- Fix query by email in admin's project page (\#1083).
 
 # 1.27.2